+++ px-map-behavior-root.es6.js
@@ -335,18 +335,6 @@
       },
 
       /**
-       * The number used to trigger the drawn layers cleanning on Map.
-       *
-       * Drawn layers will be removed when drawLayersRemovalFactor value is changed.
-       */
-      drawLayersRemovalFactor: {
-        type: Number,
-        value: 0,
-        notify: true,
-        observer: 'shouldUpdateInst'
-      },
-
-      /**
        * The controls setting for the drawing toolbar.
        *
        * See https://github.com/geoman-io/leaflet-geoman
@@ -357,9 +345,11 @@
           return {
             drawMarker: false,
             drawCircleMarker: false,
+            drawRectangle: false,
             drawPolyline: false,
             rotateMode: false,
-            drawText: false
+            drawText: false,
+            editControls: false
           };
         }
       },
@@ -407,9 +397,6 @@
       if(lastOptions.drawToolbarLanguage !== nextOptions.drawToolbarLanguage) {
         this.elementInst.pm.setLang(nextOptions.drawToolbarLanguage);
       }
-      if(lastOptions.drawLayersRemovalFactor !== nextOptions.drawLayersRemovalFactor) {
-        this._removeDrawnLayers();
-      }
 
       PxMapBehavior.LeafletRootImpl.updateInst.call(this, lastOptions, nextOptions);
     },
@@ -420,7 +407,6 @@
       options.showDrawToolbar = this.showDrawToolbar;
       options.drawToolbarLanguage = this.drawToolbarLanguage;
       options.drawControls = this.drawControls;
-      options.drawLayersRemovalFactor = this.drawLayersRemovalFactor;
 
       return options;
     },
@@ -448,15 +434,17 @@
       const layerCreatedFn = this._handleShapeLayerCreated.bind(this);
       const layerDrawEndFn = this._handleShapeLayerDrawEnd.bind(this);
       const layerRemovedFn = this._handleShapeLayerRemoved.bind(this);
+      const removalModeToggledFn = this._handleRemovalModeToggled.bind(this);
       this.bindEvents({
         'pm:create' : layerCreatedFn,
         'pm:drawend' : layerDrawEndFn,
-        'pm:remove' : layerRemovedFn
+        'pm:remove' : layerRemovedFn,
+        'pm:globalremovalmodetoggled': removalModeToggledFn
       });
 
       // creates custom button
       const drawFinishedFn = this._handleDrawFinished.bind(this);
-      this.elementInst.pm.Toolbar.createCustomControl({
+      this.finishBtn = this.elementInst.pm.Toolbar.createCustomControl({
         name: 'customFinish',
         block: 'custom',
         title: this.drawFinishTitle,
@@ -468,6 +456,22 @@
 
       // set language
       this.elementInst.pm.setLang(language);
+      // disable controls' submenu/actions
+      this._disableActionsForControls();
+    },
+
+    /**
+     * Disable/remove the submenu/actions for controls
+     */
+    _disableActionsForControls() {
+      this.elementInst.pm.Toolbar.changeActionsOfControl('Rectangle', []);
+      this.elementInst.pm.Toolbar.changeActionsOfControl('Polygon', []);
+      this.elementInst.pm.Toolbar.changeActionsOfControl('Circle', []);
+
+      this.elementInst.pm.Toolbar.changeActionsOfControl('Cut', []);
+      this.elementInst.pm.Toolbar.changeActionsOfControl('editMode', []);
+      this.elementInst.pm.Toolbar.changeActionsOfControl('dragMode', []);
+      this.elementInst.pm.Toolbar.changeActionsOfControl('removalMode', []);
     },
 
     /**
@@ -483,9 +487,15 @@
      * Remove the drawn layers
      */
     _removeDrawnLayers() {
+      // remove all drawn layers
       const layers = this.elementInst.pm.getGeomanDrawLayers();
       layers.forEach(layer => layer.remove());
+      // remove the drawing layer
+      this.elementInst.pm.disableDraw();
+      // disable the Finish button
       this._toggleFinishBtnEnalbeState(false);
+
+      this.fire('px-map-drawn-layer-changed', { hasDrawnLayers: false });
     },
 
     /**
@@ -502,6 +512,28 @@
      */
     _toggleFinishBtnEnalbeState(enable) {
       this.elementInst.pm.Toolbar.setButtonDisabled('customFinish', !enable);
+      if (this.finishBtn) {
+        const btnContainer = this.finishBtn.getContainer();
+        const btnIcon = this.finishBtn.buttonsDomNode.getElementsByClassName('control-icon')[0];
+        if (enable) {
+          btnContainer.classList.add('blue-border');
+          btnIcon.classList.add('leaflet-pm-icon-checkmark-blue');
+        } else {
+          btnContainer.classList.remove('blue-border');
+          btnIcon.classList.remove('leaflet-pm-icon-checkmark-blue');
+        }
+      }
+    },
+
+    /**
+     * Enable/Disable the shape buttons
+     *
+     * @param {*} enable
+     */
+    _toggleShapeBtnsEnableState(enable) {
+      this.elementInst.pm.Toolbar.setButtonDisabled('Rectangle', !enable);
+      this.elementInst.pm.Toolbar.setButtonDisabled('Polygon', !enable);
+      this.elementInst.pm.Toolbar.setButtonDisabled('Circle', !enable);
     },
 
     /**
@@ -545,10 +577,15 @@
 
     /**
      * Handle when the layer or shape is draw ended.
+     *
      * @param {*} e
      */
     _handleShapeLayerDrawEnd(e) {
-      this._toggleFinishBtnEnalbeState(true);
+      // drawend event is triggerred as well when disable the draw for the first shape/layer
+      // can't use hasLayers = true directly here
+      const hasLayers = this._hasDrawnLayers();
+      this._toggleFinishBtnEnalbeState(hasLayers);
+      this.fire('px-map-drawn-layer-changed', { hasDrawnLayers: hasLayers });
     },
 
     /**
@@ -556,7 +593,9 @@
      * @param {*} e
      */
     _handleShapeLayerRemoved(e) {
-      this._toggleFinishBtnEnalbeState(this._hasDrawnLayers());
+      const hasLayers = this._hasDrawnLayers();
+      this._toggleFinishBtnEnalbeState(hasLayers);
+      this.fire('px-map-drawn-layer-changed', { hasDrawnLayers: hasLayers });
     },
 
     /**
@@ -579,6 +618,16 @@
 
       this.fire('px-map-draw-finished', group.toGeoJSON());
       this._toggleFinishBtnEnalbeState(false);
+      this.elementInst.pm.disableGlobalRemovalMode();
+    },
+
+    /**
+     * Handle the event when removal mode is toggled.
+     *
+     * @param {*} evt
+     */
+    _handleRemovalModeToggled(evt) {
+      this.fire('px-map-removal-mode-changed', {removalEnabled: this.elementInst.pm.globalRemovalModeEnabled()});
     }
     /**
      * Fired when user clicking on the custom Finish button.
@@ -586,6 +635,22 @@
      *   * {Object} detail - Contains the event details, the feature colletion of drawn polygons.
      *
      * @event px-map-draw-finished
+     */
+
+    /**
+     * Fired when the layer is added/removed.
+     *
+     *   * {Object} detail - Contains the event details - flag to indicate still has drawn layers or not.
+     *
+     * @event px-map-drawn-layer-changed
+     */
+
+    /**
+     * Fired when the removal mode is toggled.
+     *
+     *   * {Object} detail - Contains the event details - removal is enable or not.
+     *
+     * @event px-map-removal-mode-changed
      */
   };